requirements.txt
torch>=2.2.0
numpy>=1.24.0
PyYAML>=6.0.0
tqdm>=4.66.0
sentencepiece>=0.2.0
datasets>=2.19.0
